requirements.txt
ansible-core
molecule
libvirt-python
pytest
lxml
netaddr
passlib
